Revert "Style book: create static categories (#65430)"
This reverts commit 6d8f2a8cb7d9da8061a495caecd3e79a98cb2309.
