+++ test/e2e/specs/site-editor/style-book.spec.js
@@ -43,6 +43,9 @@ test.describe( 'Style Book', () => {
 		await expect( page.locator( 'role=tab[name="Text"i]' ) ).toBeVisible();
 		await expect( page.locator( 'role=tab[name="Media"i]' ) ).toBeVisible();
 		await expect(
+			page.locator( 'role=tab[name="Design"i]' )
+		).toBeVisible();
+		await expect(
 			page.locator( 'role=tab[name="Widgets"i]' )
 		).toBeVisible();
 		await expect( page.locator( 'role=tab[name="Theme"i]' ) ).toBeVisible();
